feature(DEV-10629): add usage of scoped `<EmotionCacheProvider/>` to `<MoniteStyleProvider/>` component
