+++ packages/sdk-react/src/core/context/MoniteThemeProvider.ts
@@ -1,6 +1,7 @@
 import { createContext, useContext } from 'react';
 
-import { Theme } from '@mui/material';
+import { ScopedCssBaselineContainerClassName } from '@/components/ContainerCssBaseline';
+import { createTheme, Theme, ThemeOptions } from '@mui/material';
 
 export const MoniteThemeContext = createContext<Theme | undefined>(undefined);
 
@@ -14,3 +15,40 @@ export function useMoniteThemeContext(): Theme {
 
   return moniteThemeContext;
 }
+
+/**
+ * Create a theme with default component's `defaultProps`
+ */
+export const createThemeWithDefaults = (
+  theme: Theme | ThemeOptions | undefined
+) =>
+  createTheme(theme, {
+    components: {
+      MuiMenu: {
+        defaultProps: {
+          classes: {
+            root: ScopedCssBaselineContainerClassName,
+          },
+        },
+      },
+      MuiModal: {
+        defaultProps: {
+          classes: {
+            root: ScopedCssBaselineContainerClassName,
+          },
+        },
+      },
+      MuiPopper: {
+        defaultProps: {
+          componentsProps: {
+            root: { className: ScopedCssBaselineContainerClassName },
+          },
+        },
+      },
+      MuiAutocomplete: {
+        defaultProps: {
+          classes: { popper: ScopedCssBaselineContainerClassName },
+        },
+      },
+    },
+  });
